Translating: number-parsing into conversion-de-cadena-a-numero.
