+++ examples/conversion-de-cadena-a-numero/number-parsing.hash
