+++ examples/conversion-de-cadena-a-numero/number-parsing.sh
@@ -1,0 +1,9 @@
+$ go run number-parsing.go
+1.234
+123
+456
+789
+135
+strconv.ParseInt: parsing "wat": invalid syntax
+
+# A continuación veremos otra tarea común de análisis de cadenas: URLs.
