--- examples/number-parsing/number-parsing.sh
@@ -1,9 +1,0 @@
-$ go run number-parsing.go 
-1.234
-123
-456
-789
-135
-strconv.ParseInt: parsing "wat": invalid syntax
-
-# Next we'll look at another common parsing task: URLs.
